+++ src/main.cpp
@@ -1,33 +1,35 @@
 #include <Geode/binding/CCMenuItemSpriteExtra.hpp>
-#include "PackSelectLayer.hpp"
-#include <Geode/modify/MenuLayer.hpp>
+#include "PackSelectPopup.hpp"
+#include <Geode/modify/OptionsLayer.hpp>
 #include <Geode/modify/IDManager.hpp>
 #include <Geode/ui/BasedButtonSprite.hpp>
 
 using namespace geode::prelude;
 
-class $modify(MyMenuLayer, MenuLayer) {
-    bool init() {
-        if (!MenuLayer::init())
-            return false;
+class $modify(MyOptionsLayer, OptionsLayer) {
 
-        NodeIDs::provideFor(this);
+    void customSetup() {
+        OptionsLayer::customSetup();
+        if (CCNode* optionsMenu = m_mainLayer->getChildByID("options-menu")) {
+            if (AxisLayout* layout = typeinfo_cast<AxisLayout*>(optionsMenu->getLayout())) {
+                layout->ignoreInvisibleChildren(true);
+            }
+            
+            ButtonSprite* buttonSprite = ButtonSprite::create("Textures", 130, true, "goldFont.fnt", "GJ_button_01.png", 30, 1);
 
-        auto menu = this->getChildByID("right-side-menu");
+            CCMenuItemSpriteExtra* textureLoaderBtn = CCMenuItemSpriteExtra::create(buttonSprite, this, menu_selector(MyOptionsLayer::onTextureLdr));
 
-        auto button = CCMenuItemSpriteExtra::create(
-            CCSprite::createWithSpriteFrameName("gj_folderBtn_001.png"),
-            this, menu_selector(MyMenuLayer::onTextureLdr)
-        );
-        button->setID("texture-loader-button"_spr);
-        menu->addChild(button);
-        menu->updateLayout();
+            if (CCNode* howToPlayBtn = optionsMenu->getChildByID("how-to-play-button")) {
+                howToPlayBtn->setVisible(false);
+                optionsMenu->insertBefore(textureLoaderBtn, howToPlayBtn);
+            }
 
-        return true;
+            optionsMenu->updateLayout();
+        }
     }
 
     void onTextureLdr(CCObject*) {
-        PackSelectLayer::scene();
+        PackSelectPopup::create()->show();
     }
 };
